Removing single and error styling
