--- eq-author/src/App/page/Design/CalculatedSummaryPageEditor/old/AnswerSelector.js
@@ -1,329 +1,0 @@
-import React, { useState, useMemo } from "react";
-import PropTypes from "prop-types";
-import styled, { css } from "styled-components";
-import { propType } from "graphql-anywhere";
-import gql from "graphql-tag";
-
-import { NUMBER, CURRENCY, UNIT, PERCENTAGE } from "constants/answer-types";
-import { calculatedSummaryErrors } from "constants/validationMessages";
-import { colors } from "constants/theme";
-
-import { MenuItemType } from "components/ContentPickerv2/Menu";
-import AnswerPicker from "components/AnswerPicker";
-import Button from "components/buttons/Button";
-import TextButton from "components/buttons/TextButton";
-import ValidationError from "components/ValidationError";
-import { useQuestionnaire } from "components/QuestionnaireContext";
-import AnswerPreview from "./AnswerPreview";
-
-import iconInfo from "assets/icon-info.svg";
-
-import getContentBeforeEntity from "utils/getContentBeforeEntity";
-
-const Box = styled.div`
-  border: 1px solid ${colors.borders};
-  border-radius: 3px;
-  margin-bottom: 2em;
-  overflow: hidden;
-`;
-
-const RemoveAllButton = styled(TextButton)`
-  letter-spacing: 0.05rem;
-  font-weight: bold;
-  font-size: 0.8rem;
-  margin: 0 0 0 auto;
-`;
-
-const Answers = styled.div`
-  padding: 1em;
-`;
-
-const SectionList = styled.ul`
-  list-style: none;
-  margin: 0 0 0.5em;
-  padding: 0;
-`;
-
-const SectionListItem = styled.li`
-  margin: 0;
-`;
-
-const SectionHeader = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-  margin-bottom: 0.7em;
-`;
-
-const SectionTitle = styled.div`
-  font-weight: bold;
-  font-size: 0.9em;
-  color: #807d77;
-`;
-
-const List = styled.ul`
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-flow: row wrap;
-`;
-
-const ListItem = styled.li`
-  margin: 0 0 0.5em;
-  width: 100%;
-
-  &:last-of-type {
-    margin-bottom: 0;
-  }
-`;
-
-const SelectButton = styled(Button)`
-  width: 100%;
-`;
-
-const Empty = styled.div`
-  color: ${colors.calcSumEmptyContent};
-  text-align: center;
-  padding: 1em 2em;
-
-  &::before {
-    display: block;
-    content: url(${iconInfo});
-  }
-`;
-
-const EmptyTitle = styled.h3`
-  font-size: 1em;
-  margin: 0 0 0.5em;
-`;
-
-const EmptyButton = styled(Button)`
-  padding: 0.5em 1em;
-`;
-
-const EmptyText = styled.div`
-  font-size: 0.9em;
-  margin-bottom: 1em;
-`;
-
-const ErrorButtonContainer = styled.div`
-  border: 2px solid ${colors.errorPrimary};
-  padding: 10px;
-  position: relative;
-`;
-
-const ErrorContainer = styled.div`
-  margin-left: 2em;
-`;
-
-const Chip = styled(MenuItemType)`
-  color: ${colors.text};
-  float: right;
-`;
-
-const Text = styled.p`
-  max-width: 30em;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  float: left;
-  margin: 0;
-`;
-
-export const ErrorContext = styled.div`
-  position: relative;
-
-  ${(props) =>
-    props.isInvalid &&
-    css`
-      border: 2px solid ${colors.errorPrimary};
-      padding: 1em;
-    `}
-`;
-
-const renderValidationError = (message) => (
-  <ValidationError>{message}</ValidationError>
-);
-
-const renderEmptyState = (handlePickerOpen, availableSummaryAnswers) => (
-  <>
-    <Empty>
-      <EmptyTitle>{`No answers ${
-        availableSummaryAnswers.length ? "selected" : "available"
-      }`}</EmptyTitle>
-      <EmptyText>
-        {availableSummaryAnswers.length
-          ? "Select an answer using the button below."
-          : "There are no answers to provide a calculated summary."}
-      </EmptyText>
-      <ErrorButtonContainer>
-        <EmptyButton
-          small
-          onClick={handlePickerOpen}
-          data-test="answer-selector-empty"
-          disabled={!availableSummaryAnswers.length}
-        >
-          Select an answer
-        </EmptyButton>
-      </ErrorButtonContainer>
-    </Empty>
-    <ErrorContainer>
-      {renderValidationError(calculatedSummaryErrors.ERR_NO_ANSWERS)}
-    </ErrorContainer>
-  </>
-);
-
-export const filterAvailableAnswers = (entity) =>
-  [CURRENCY, UNIT, PERCENTAGE, NUMBER].includes(entity.type) ? entity : [];
-
-const AnswerList = ({ answers, validationErrorCode, handleRemoveAnswers }) => (
-  <>
-    <ErrorContext isInvalid={validationErrorCode}>
-      <List>
-        {answers.map((data) => (
-          <ListItem key={data.id}>
-            <AnswerPreview onRemove={() => handleRemoveAnswers([data])}>
-              <Text>{data.displayName}</Text>
-              {data.properties.unit && (
-                <Chip key={data.properties.unit}>{data.properties.unit}</Chip>
-              )}
-              <Chip key={data.type}>{data.type}</Chip>
-            </AnswerPreview>
-          </ListItem>
-        ))}
-      </List>
-    </ErrorContext>
-    {validationErrorCode &&
-      renderValidationError(calculatedSummaryErrors[validationErrorCode])}
-  </>
-);
-
-const renderChosenAnswers = (
-  answers,
-  answerType,
-  section,
-  validationErrorCode,
-  handlePickerOpen,
-  handleRemoveAnswers
-) => (
-  <>
-    <SectionList>
-      <SectionListItem key={section.id}>
-        <SectionHeader>
-          <SectionTitle>
-            {answerType} answers in {section.displayName}
-          </SectionTitle>
-          <RemoveAllButton
-            data-test="remove-all"
-            onClick={() => handleRemoveAnswers(answers)}
-          >
-            Remove all
-          </RemoveAllButton>
-        </SectionHeader>
-        <AnswerList
-          answers={answers}
-          validationErrorCode={validationErrorCode}
-          handleRemoveAnswers={handleRemoveAnswers}
-        />
-      </SectionListItem>
-    </SectionList>
-    <SelectButton
-      variant="secondary"
-      onClick={handlePickerOpen}
-      data-test="answer-selector"
-    >
-      Select another {answerType?.toLowerCase() || ""} answer
-    </SelectButton>
-  </>
-);
-
-export const AnswerSelector = ({ onUpdateCalculatedSummaryPage, page }) => {
-  const [showPicker, setShowPicker] = useState(false);
-  const { questionnaire } = useQuestionnaire();
-  const { summaryAnswers, validationErrorInfo } = page;
-
-  const availableSummaryAnswers = useMemo(
-    () =>
-      (
-        questionnaire &&
-        getContentBeforeEntity({
-          questionnaire,
-          id: page.id,
-          preprocessAnswers: filterAvailableAnswers,
-        })
-      )?.filter((section) => section.id === page.section.id) || [],
-    [questionnaire, page.id, page.section.id]
-  );
-
-  const handleRemoveAnswers = (answers) =>
-    onUpdateCalculatedSummaryPage({
-      id: page.id,
-      summaryAnswers: summaryAnswers.filter(
-        ({ id }) => !answers.find((answer) => answer.id === id)
-      ),
-    });
-
-  const handlePickerOpen = () => setShowPicker(true);
-  const handlePickerClose = () => setShowPicker(false);
-
-  const handlePickerSubmit = (answers) => {
-    handlePickerClose();
-    onUpdateCalculatedSummaryPage({
-      id: page.id,
-      summaryAnswers: answers,
-    });
-  };
-
-  return (
-    <Box>
-      <Answers>
-        {summaryAnswers.length
-          ? renderChosenAnswers(
-              summaryAnswers,
-              summaryAnswers?.[0]?.type,
-              page.section,
-              validationErrorInfo?.errors[0]?.errorCode,
-              handlePickerOpen,
-              handleRemoveAnswers
-            )
-          : renderEmptyState(handlePickerOpen, availableSummaryAnswers)}
-        <AnswerPicker
-          isOpen={showPicker}
-          onClose={handlePickerClose}
-          onSubmit={handlePickerSubmit}
-          startingSelectedAnswers={summaryAnswers}
-          data={availableSummaryAnswers}
-          title="Select one or more answer"
-          showTypes
-        />
-      </Answers>
-    </Box>
-  );
-};
-
-// AnswerSelector.fragments = {
-//   AnswerSelector: gql`
-//     fragment AnswerSelector on CalculatedSummaryPage {
-//       id
-//       section {
-//         id
-//         displayName
-//       }
-//       summaryAnswers {
-//         id
-//         displayName
-//         type
-//         properties
-//       }
-//     }
-//   `,
-// };
-
-AnswerSelector.propTypes = {
-  onUpdateCalculatedSummaryPage: PropTypes.func.isRequired,
-  page: propType(AnswerSelector.fragments.AnswerSelector),
-  getValidationError: PropTypes.func.isRequired,
-};
-
-export default AnswerSelector;
